Move mockApi to lower directory (within components).
