+++ src/uniprotkb/components/__mocks__/fasta.json
