+++ src/uniprotkb/components/__mocks__/noResults.json
