+++ src/uniprotkb/components/__mocks__/proteinNamesUIData.json
