+++ src/uniprotkb/components/__mocks__/swissprotEntry.json
