+++ src/uniprotkb/components/entry/__tests__/ProteinOverview.spec.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { render } from '@testing-library/react';
 import ProteinOverview from '../../protein-data-views/ProteinOverviewView';
-import ProteinNamesUIData from '../../../__mocks__/proteinNamesUIData.json';
+import ProteinNamesUIData from '../../__mocks__/proteinNamesUIData.json';
 import EntrySection from '../../../types/entrySection';
 
 describe('ProteinOverview component', () => {
