+++ src/uniprotkb/components/protein-data-views/__tests__/ProteinNamesView.spec.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { render } from '@testing-library/react';
 import EntryProteinNames from '../ProteinNamesView';
-import ProteinNamesUIData from '../../../__mocks__/proteinNamesUIData.json';
+import ProteinNamesUIData from '../../__mocks__/proteinNamesUIData.json';
 
 describe('ProteinNames', () => {
   test('should render protein_name', () => {
